departure_prediction/requirements.txt
# ============================================
# AI-Enabled Flight Departure Time Predictor
# ============================================

# Core ML/DL
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# API & Web
requests>=2.31.0
python-dotenv>=1.0.0

# Image Processing (Ticket OCR)
Pillow>=10.0.0

# Jupyter Notebooks (for model training)
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# XGBoost (Alternative predictor)
xgboost>=2.0.0

# Utilities
tqdm>=4.65.0

# ============================================
# External Services (require API keys)
# ============================================
# - Ollama (local LLM): Install separately from ollama.ai
# - Google Maps API: For Routes & Weather
# - AviationStack API: For real-time flight data

